Extracting common base class for data extraction processes for L1B/2A, L2B/C and AUX data. Improving AUX process to enable async processes and netCDF download.
